Refactor: libcrmcommon: Add logging_internal.h
This moves some internal definitions from include/crm/common/internal.h
to a more specific header file.
