+++ include/crm_internal.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2006-2020 the Pacemaker project contributors
+ * Copyright 2006-2021 the Pacemaker project contributors
  *
  * The version control history for this file may have further details.
  *
@@ -30,6 +30,7 @@
 
 #  include <crm/lrmd.h>
 #  include <crm/common/logging.h>
+#  include <crm/common/logging_internal.h>
 #  include <crm/common/ipc_internal.h>
 #  include <crm/common/options_internal.h>
 #  include <crm/common/internal.h>
